pyproject.toml
[project]
name = "file-transfer"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django>=5.2",
    "djangorestframework>=3.16.0",
    "paramiko>=3.5.1",
    "tqdm>=4.67.1",
]
